Split clusters tests
