+++ e2e/atlas/clusters_flags_test.go
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//go:build e2e || (atlas && clusters && !m0)
+//go:build e2e || (atlas && clusters && flags)
 
 package atlas_test
 
@@ -23,7 +23,6 @@ import (
 	"testing"
 
 	"github.com/mongodb/mongocli/e2e"
-	"github.com/mongodb/mongocli/internal/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.mongodb.org/atlas/mongodbatlas"
@@ -226,174 +225,6 @@ func TestClustersFlags(t *testing.T) {
 			clustersEntity,
 			"watch",
 			clusterName,
-			"--projectId", g.projectID,
-		)
-		cmd.Env = os.Environ()
-		// this command will fail with 404 once the cluster is deleted
-		// we just need to wait for this to close the project
-		resp, _ := cmd.CombinedOutput()
-		t.Log(string(resp))
-	})
-}
-
-func TestClustersFile(t *testing.T) {
-	g := newClusterGenerator(t)
-	g.generateProject("clustersFile")
-
-	cliPath, err := e2e.Bin()
-	req := require.New(t)
-	req.NoError(err)
-
-	clusterFileName, err := RandClusterName()
-	req.NoError(err)
-
-	clusterFile := "create_cluster_test.json"
-	if service := os.Getenv("MCLI_SERVICE"); service == config.CloudGovService {
-		clusterFile = "create_cluster_gov_test.json"
-	}
-
-	t.Run("Create via file", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			atlasEntity,
-			clustersEntity,
-			"create",
-			clusterFileName,
-			"--file", clusterFile,
-			"--projectId", g.projectID,
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		req.NoError(err, string(resp))
-
-		var cluster mongodbatlas.AdvancedCluster
-		err = json.Unmarshal(resp, &cluster)
-		req.NoError(err)
-
-		ensureCluster(t, &cluster, clusterFileName, e2eMDBVer, 30)
-	})
-
-	t.Run("Watch", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			atlasEntity,
-			clustersEntity,
-			"watch",
-			"--projectId", g.projectID,
-			clusterFileName,
-		)
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		req.NoError(err, string(resp))
-
-		a := assert.New(t)
-		a.Contains(string(resp), "Cluster available")
-	})
-
-	t.Run("Update via file", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			atlasEntity,
-			clustersEntity,
-			"update",
-			clusterFileName,
-			"--file=update_cluster_test.json",
-			"--projectId", g.projectID,
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		req.NoError(err, string(resp))
-
-		var cluster mongodbatlas.AdvancedCluster
-		err = json.Unmarshal(resp, &cluster)
-		req.NoError(err)
-
-		ensureCluster(t, &cluster, clusterFileName, e2eMDBVer, 40)
-	})
-
-	t.Run("Delete file creation", func(t *testing.T) {
-		cmd := exec.Command(cliPath, atlasEntity, clustersEntity, "delete", clusterFileName, "--projectId", g.projectID, "--force")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		req.NoError(err, string(resp))
-
-		expected := fmt.Sprintf("Cluster '%s' deleted\n", clusterFileName)
-		a := assert.New(t)
-		a.Equal(expected, string(resp))
-	})
-
-	t.Run("Watch deletion", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			atlasEntity,
-			clustersEntity,
-			"watch",
-			clusterFileName,
-			"--projectId", g.projectID,
-		)
-		cmd.Env = os.Environ()
-		// this command will fail with 404 once the cluster is deleted
-		// we just need to wait for this to close the project
-		resp, _ := cmd.CombinedOutput()
-		t.Log(string(resp))
-	})
-}
-
-func TestShardedCluster(t *testing.T) {
-	g := newClusterGenerator(t)
-	g.generateProject("shardedClusters")
-
-	cliPath, err := e2e.Bin()
-	a := assert.New(t)
-	req := require.New(t)
-	req.NoError(err)
-
-	shardedClusterName, err := RandClusterName()
-	req.NoError(err)
-
-	region, err := g.newAvailableRegion(tierM30, e2eClusterProvider)
-	req.NoError(err)
-
-	t.Run("Create sharded cluster", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			atlasEntity,
-			clustersEntity,
-			"create",
-			shardedClusterName,
-			"--region", region,
-			"--type=SHARDED",
-			"--shards=2",
-			"--members=3",
-			"--tier", tierM30,
-			"--provider", e2eClusterProvider,
-			"--mdbVersion=4.2",
-			"--diskSizeGB", diskSizeGB30,
-			"--projectId", g.projectID,
-			"-o=json")
-
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		req.NoError(err, string(resp))
-
-		var cluster mongodbatlas.AdvancedCluster
-		err = json.Unmarshal(resp, &cluster)
-		req.NoError(err)
-
-		ensureCluster(t, &cluster, shardedClusterName, "4.2", 30)
-	})
-
-	t.Run("Delete sharded cluster", func(t *testing.T) {
-		cmd := exec.Command(cliPath, atlasEntity, clustersEntity, "delete", shardedClusterName, "--projectId", g.projectID, "--force")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		req.NoError(err, string(resp))
-
-		expected := fmt.Sprintf("Cluster '%s' deleted\n", shardedClusterName)
-		a.Equal(expected, string(resp))
-	})
-
-	t.Run("Watch deletion", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			atlasEntity,
-			clustersEntity,
-			"watch",
-			shardedClusterName,
 			"--projectId", g.projectID,
 		)
 		cmd.Env = os.Environ()
