+++ e2e/atlas/clusters_sharded_test.go
@@ -1,0 +1,98 @@
+// Copyright 2020 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//go:build e2e || (atlas && clusters && sharded)
+
+package atlas_test
+
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+	"os/exec"
+	"testing"
+
+	"github.com/mongodb/mongocli/e2e"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"go.mongodb.org/atlas/mongodbatlas"
+)
+
+func TestShardedCluster(t *testing.T) {
+	g := newClusterGenerator(t)
+	g.generateProject("shardedClusters")
+
+	cliPath, err := e2e.Bin()
+	a := assert.New(t)
+	req := require.New(t)
+	req.NoError(err)
+
+	shardedClusterName, err := RandClusterName()
+	req.NoError(err)
+
+	region, err := g.newAvailableRegion(tierM30, e2eClusterProvider)
+	req.NoError(err)
+
+	t.Run("Create sharded cluster", func(t *testing.T) {
+		cmd := exec.Command(cliPath,
+			atlasEntity,
+			clustersEntity,
+			"create",
+			shardedClusterName,
+			"--region", region,
+			"--type=SHARDED",
+			"--shards=2",
+			"--members=3",
+			"--tier", tierM30,
+			"--provider", e2eClusterProvider,
+			"--mdbVersion=4.2",
+			"--diskSizeGB", diskSizeGB30,
+			"--projectId", g.projectID,
+			"-o=json")
+
+		cmd.Env = os.Environ()
+		resp, err := cmd.CombinedOutput()
+		req.NoError(err, string(resp))
+
+		var cluster mongodbatlas.AdvancedCluster
+		err = json.Unmarshal(resp, &cluster)
+		req.NoError(err)
+
+		ensureCluster(t, &cluster, shardedClusterName, "4.2", 30)
+	})
+
+	t.Run("Delete sharded cluster", func(t *testing.T) {
+		cmd := exec.Command(cliPath, atlasEntity, clustersEntity, "delete", shardedClusterName, "--projectId", g.projectID, "--force")
+		cmd.Env = os.Environ()
+		resp, err := cmd.CombinedOutput()
+		req.NoError(err, string(resp))
+
+		expected := fmt.Sprintf("Cluster '%s' deleted\n", shardedClusterName)
+		a.Equal(expected, string(resp))
+	})
+
+	t.Run("Watch deletion", func(t *testing.T) {
+		cmd := exec.Command(cliPath,
+			atlasEntity,
+			clustersEntity,
+			"watch",
+			shardedClusterName,
+			"--projectId", g.projectID,
+		)
+		cmd.Env = os.Environ()
+		// this command will fail with 404 once the cluster is deleted
+		// we just need to wait for this to close the project
+		resp, _ := cmd.CombinedOutput()
+		t.Log(string(resp))
+	})
+}
